requirements.txt
Pillow==10.0.1
streamlit==1.28.2
